MEN-56: Move logging framework to separate repositories.
This way both server and client can use it.
