+++ Godeps/Godeps.json
@@ -1,9 +1,6 @@
 {
-	"ImportPath": "mender",
+	"ImportPath": "github.com/mendersoftware/mender",
 	"GoVersion": "go1.5",
-	"Packages": [
-		"github.com/Sirupsen/logrus"
-	],
 	"Deps": [
 		{
 			"ImportPath": "github.com/Sirupsen/logrus",
@@ -11,21 +8,16 @@
 			"Rev": "be52937128b38f1d99787bb476c789e2af1147f1"
 		},
 		{
-			"ImportPath": "github.com/davecgh/go-spew/spew",
-			"Rev": "5215b55f46b2b919f50a1df0eaa5886afe4e3b3d"
+			"ImportPath": "github.com/mendersoftware/log",
+			"Rev": "57b90bcceb46b1f8a8b60c83f177a2d40bd7b44d"
 		},
 		{
-			"ImportPath": "github.com/pmezard/go-difflib/difflib",
-			"Rev": "792786c7400a136282c1664665ae0a8db921c6c2"
+			"ImportPath": "github.com/mendersoftware/mendertesting",
+			"Rev": "185b0658c0984b85ed862f55c77a62ba8401fca7"
 		},
 		{
-			"ImportPath": "github.com/stretchr/testify/assert",
-			"Comment": "v1.1.3-2-g9f9027f",
-			"Rev": "9f9027faeb0dad515336ed2f28317f9f8f527ab4"
-		},
-		{
-			"ImportPath": "golang.org/x/sys/unix",
-			"Rev": "50c6bc5e4292a1d4e65c6e9be5f53be28bcbe28e"
+			"ImportPath": "github.com/mendersoftware/scopestack",
+			"Rev": "feb2f3f2bec86824d6ac1963e934c790c51b2776"
 		}
 	]
 }
