+++ Godeps/_workspace/src/github.com/mendersoftware/mendertesting/mendertesting.go
