+++ Godeps/_workspace/src/github.com/mendersoftware/scopestack/scope_stack.go
